flexmind-ros1-release-1.0.0/requirements.txt
absl-py==2.3.1
actionlib==1.14.3
angles==1.9.14
apt-clone==0.2.1
apturl==0.5.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
blinker==1.4
bondpy==1.8.7
Brlapi==0.7.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
distro==1.4.0
distro-info===0.23ubuntu1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.23
docstring_parser==0.17.0
docutils==0.16
duplicity==0.8.12.0
dynamic-reconfigure==1.7.6
empy==3.3.2
entrypoints==0.3
etils==1.3.0
eval_type_backport==0.2.2
fasteners==0.14.1
future==0.18.2
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
glfw==2.10.0
graphsurgeon==0.4.6
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==16.10.1
interactive-markers==1.12.2
Jetson.GPIO==2.1.1
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
keyring==18.0.1
kiwisolver==1.0.1
labmaze==1.0.6
language-selector==0.1
laser-geometry==1.6.8
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==6.0.1
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==1.1.0
matplotlib==3.1.2
mcap==1.3.0
mcap-ros1-support==0.7.3
mdurl==0.1.2
message-filters==1.17.4
monotonic==1.5
mujoco==3.2.3
netifaces==0.10.4
nose==1.3.7
numpy==1.17.4
numpy-quaternion==2023.0.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
onnx-graphsurgeon==0.3.12
PAM==0.4.2
pandas==0.25.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pkg_resources==0.0.0
protobuf==5.29.5
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
Pygments==2.19.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.10
pyOpenSSL==19.0.0
pyparsing==3.1.4
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyrealsense2==2.55.1.6486
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.1
python-dateutil==2.7.3
python-dbusmock==0.19
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python-qt-binding==0.4.6
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
Quaternion==3.5.2.post4
requests==2.22.0
requests-unixsocket==0.2.0
resource-retriever==1.12.10
rich==14.1.0
roman==2.0.0
rosapi==0.11.17
rosbag==1.17.4
rosboost-cfg==1.15.10
rosbridge-library==0.11.17
rosbridge-server==0.11.17
rosclean==1.15.10
roscreate==1.15.10
rosdep-modules==0.25.1
rosdistro-modules==1.0.1
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rviz==1.14.26
scipy==1.3.3
SecretStorage==2.3.1
sensor-msgs==1.13.2
service-identity==18.1.0
shtab==1.7.2
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
systemd-python==234
tensorrt==8.5.2.2
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
topic-tools==1.17.4
tornado==5.1.1
tqdm==4.67.1
transitions==0.9.3
Twisted==18.9.0
txaio==2.10.0
typeguard==4.4.0
typing_extensions==4.13.2
tyro==0.9.32
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uff==0.6.9
urllib3==1.25.8
urwid==2.0.1
wadllib==1.3.3
wsaccel==0.6.2
xacro==1.14.20
xkit==0.0.0
zipp==3.20.2
zope.interface==4.7.1
zstandard==0.23.0
mcap==1.3.0
mcap-ros1-support==0.7.3